Improve performance testing scripts

+++ testing/geekbench/download_geekbench.sh
@@ -2,9 +2,13 @@
 
 SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )"
 
-VERSIONS=("6.5.0" "5.4.5" "4.4.4" "3.4.4" "2.4.3")
+# https://www.geekbench.com/ai/download/linux/
+# https://www.geekbench.com/download/linux/
+# https://www.geekbench.com/legacy/
+
+VERSIONS=("AI-1.6.0" "-6.5.0" "-5.5.1" "-4.4.4" "-3.4.4" "-2.4.3")
 for VERSION in "${VERSIONS[@]}"; do
-  FOLDER_NAME="Geekbench-${VERSION}-Linux"
+  FOLDER_NAME="Geekbench${VERSION}-Linux"
   if [ -d "${SCRIPT_DIR}/${FOLDER_NAME}" ]; then
     echo "Geekbench ${VERSION} seems to be already downloaded."
   else
@@ -12,8 +16,10 @@ for VERSION in "${VERSIONS[@]}"; do
     FILENAME="${FOLDER_NAME}.tar.gz"
     wget "https://cdn.geekbench.com/${FILENAME}" -O "${SCRIPT_DIR}/${FILENAME}"
     tar -xzf "${FILENAME}"
-    rm "${FILENAME}"
+    # rm "${FILENAME}"
   fi
 done
-mv -f "${SCRIPT_DIR}/dist/"* "${SCRIPT_DIR}"
-rm -rf "${SCRIPT_DIR}/dist"
+if [ -d "${SCRIPT_DIR}/dist" ]; then
+  mv -f "${SCRIPT_DIR}/dist/"* "${SCRIPT_DIR}"
+  rm -rf "${SCRIPT_DIR}/dist"
+fi
